Cambios de nombre de los scripts Barrel y OtherBarrel, a FirstBarrel y SecondBarrel, además se agrego una super clase Barrel de la cual heredan algunos atributos y metodos las clases FirstBarrel y SecondBarrel
