+++ DonkeyKongPVJs/Assets/PlatformCollisionHandler.cs
@@ -6,9 +6,9 @@ public class PlatformCollisionHandler : MonoBehaviour, IBarrelCollisionHandler
 {
     [SerializeField] private float forceMultiplier = 1f;
 
-    public void HandleCollision(Barrel barrel, Collision2D collision)
+    public void HandleCollision(FirtsBarrel firtsBarrel, Collision2D collision)
     {
-        Rigidbody2D rb = barrel.GetComponent<Rigidbody2D>();
+        Rigidbody2D rb = firtsBarrel.GetComponent<Rigidbody2D>();
         if (rb != null)
         {
             rb.AddForce(collision.transform.right * forceMultiplier, ForceMode2D.Impulse);
