+++ DonkeyKongPVJs/Assets/Scripts/FirtsBarrel.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0b8456125faebba4da2c38f915ef4d6a
+guid: cd6914d88cc1dfd4696d90bd3af7a613
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
